Create onDragEnd function in utils file and use it in components which uses drag and drop
